feat(trace) add trace scheduler
